add promise and fix old code

+++ bot.js
@@ -85,7 +85,7 @@ client.on(`message`, async message => {
     //trivia
     var triviarandom = Math.floor(Math.random()*14);
     if((triviarandom == 0 && message.channel.name == "general" && !questionActive) || (message.author.id == '417439359868862465' && message.content == "triggertrivia" && message.channel.name == "general")) { // MAKE IT SO THAT THIS CAN'T TRIGGER DURING A QUESTION
-        setTimeout(function(){message.delete()}, 1000)
+        if(message.author.id == '417439359868862465' && message.content == "triggertrivia") setTimeout(function(){message.delete()}, 1000);
         questionActive = true;
         trivia.execute(message.channel, triviaquestions, Discord, obols, fs, questionActive, client).then(() => {
             questionActive = false;
